feat: changes implements in all files App

+++ App/crud.py
@@ -34,7 +34,7 @@ def create_reservation(reservation_data):
     
     try:
         cursor.execute(
-            "SELECT * FROM reservations WHERE reservation_datetime = %s AND status = 'active'",
+            "SELECT * FROM reservations WHERE reservation_datetime = %s AND status = 'activa'",
             (reservation_data['reservation_datetime'],)
         )
         existing_reservation = cursor.fetchall()
@@ -42,8 +42,8 @@ def create_reservation(reservation_data):
             return {"error": "Ya existe una reserva para esta fecha y hora."}
         total_cost = calculate_total_cost(reservation_data['age'])
         cursor.execute(
-            "INSERT INTO reservations (customer_name, number_of_people, reservation_datetime, status, age, total_cost) VALUES (%s, %s, %s, %s, %s, %s)",
-            (reservation_data['customer_name'], reservation_data['number_of_people'], reservation_data['reservation_datetime'], reservation_data['status'], reservation_data['age'], total_cost)
+            "INSERT INTO reservations (customer_name, reservation_datetime, status, age, total_cost) VALUES (%s, %s, %s, %s, %s)",
+            (reservation_data['customer_name'], reservation_data['reservation_datetime'], 'activa', reservation_data['age'], total_cost)
         )
         connectMySQL.commit()
         new_id = cursor.lastrowid
@@ -60,25 +60,33 @@ def create_reservation(reservation_data):
     return get_reservation(new_id)
 
 
-
-def update_reservation(reservation_id: int, reservation_data):
+def update_reservation(reservation_id: int, reservation_data: dict):
     cursor = connectMySQL.cursor()
-    total_cost = calculate_total_cost(reservation_data['age'])
+    customer_name = reservation_data['customer_name']
+    reservation_datetime = reservation_data['reservation_datetime']
+    status = reservation_data['status']
+    age = int(reservation_data['age'])
+    total_cost = calculate_total_cost(age)
     cursor.execute(
-        "UPDATE reservations SET customer_name=%s, number_of_people=%s, reservation_datetime=%s, status=%s, age=%s, total_cost=%s WHERE id=%s",
-        (reservation_data['customer_name'], reservation_data['number_of_people'], reservation_data['reservation_datetime'], reservation_data['status'], reservation_data['age'], total_cost, reservation_id)
+        "UPDATE reservations SET customer_name=%s, reservation_datetime=%s, status=%s, age=%s, total_cost=%s WHERE id=%s",
+        (customer_name, reservation_datetime, status, age, total_cost, reservation_id)
     )
+    
     connectMySQL.commit()
     cursor.close()
     return get_reservation(reservation_id)
 
 def delete_reservation(reservation_id: int):
-    cursor = connectMySQL.cursor()
+    cursor = connectMySQL.cursor(dictionary=True)
+    cursor.execute("SELECT * FROM reservations WHERE id = %s", (reservation_id,))
+    reservation = cursor.fetchone()
+    if not reservation:
+        return None
     cursor.execute("DELETE FROM reservations WHERE id = %s", (reservation_id,))
     connectMySQL.commit()
-    rows_affected = cursor.rowcount
     cursor.close()
-    return rows_affected > 0
+    return reservation
+
 
 def create_user(user_data):
     cursor = connectMySQL.cursor()
